Cop-out on CUDA arrays
Program returns correct answers!

However, this was done by allocating memory and passing the height and
width values to the kernel.  This is less than ideal as it should be
possible to simply pass an array (or something along the lines of an
array) to the kernel that can be processed without using a predetermined
variable.  Will look into this option in new branch.

It is possible to simply pass integers or other values directly to a
kernel.  However this removes a number of checks (aka all the "oops" if
statements) that ensure all memory was correctly transferred to the
device.  This code will continue to use copy and place methods as is
standard in all CUDA code.