Comparison of FillArray VS Standard Nested Loop
Attempted to add parallel processing when declaring array.
Unfortunately a direct time comparison between HansenComputeGPU and
hansencompute-unidat shows that the change actually slowed the process
down.

HansenComputeGPU took  0.61s to fill array while hansencompute-unidat
took 0.00.  As such, the kernel FillArrayInt, FillArrayDouble, and
FillArrayFloat will not be used in future code.  If no application calls
for them in the next few weeks, then they will be removed to ensure that
library code is clear and concise.