Work on Implementing GPU in HansenComputeGPU
Added timing code to CudaLib and testCudaLib to see how long CUDA calls
take.  Typically these processes take about .6 seconds due to overhead
from memory allocations and memory copies between Host and Device.

I am currently working on implementing GPU processing into the
HansenComputeGPU program.  Unfortunately, the kernel was incorrectly
setup such that it added elements based on the specific image.  This
yielded return matrices with with incorrect dimensions, which in-turn
caused the return memory copies to fail.  Work is now being done to
correct this mistake and setup the kernel as a 2D system of threads.
Admittedly this system will likely max out at 512 elements but research
has indicated that the GPU should not throw an error.  Data along "y"
(row or nimage) dimension of the matrix will be summed using reduction.
The "x" (col or nmolec) dimension will be used to ensure that the proper
values are being added or subtracted from one another.