Successful Implementation of GPU and Setup for GPU Optimization
I successfully implemented the GPU code into HansenComputeGPU.cpp.  The
program now returns the same values as hansencompute-unidat.cpp.

Now that the GPU program has been implemented a quick time test was
performed.   Memory transfer to GPU took a total of 0.6 seconds for
allocation and copying.  Running Kernels and retrieving two arrays of
data took approximately 0.01 seconds.  As a result, I am now going to
attempt to optimize the GPU code using massive test arrays.

Also, the function gpuErr has been added to the ClassLib.h (will likely
change name of header file in near future to better reflect what it is
holding).  This function will replace the oops = / if (oops...)
convention that has been used so far in this code.  The main purpose of
this will to be ensure that CUDA calls are behaving as expected.