Creation of Performance Test
Created performance test to see how the GPU portion of code could be
optimized.  Ideally this project will give some insight into how CUDA
code can be manipulated to improve its performance when calculating
sums.

While code is almost ready to run, there are a few issues with the
massive 2D array that is going to be used in this test.  These issues
result in seg faults (horray!!) which are taking place around line 49 of
testCudaPerf.cpp.  Hopefully these issues will be resolved in the near
future.