Moving over to vectors
Having issues with seg fault around line 56.  For some reason the array
is declared but immediately throws a seg fault.  All data becomes
inaccessible after seg fault is thrown (as expected).

In an attempt to reduce the stress of dealing with pointers and arrays
(and prevent myself from going bald), I will be converting code such
that it uses vectors.  This is ideally the last commit that will use
arrays.  That said, HansenComputeGPU will likely not be converted such
that it uses arrays.  The theory behind this is that the GPU code
applied to vectors should apply the same way to std arrays that already
work in HansenComputeGPU.