Addition of THRUST
The changes associated with this commit involve the attempted use of the
THRUST library to successfully transfer vectors between the host and the
device.  However, this change was unsuccessful.  As such this commit
will immediately be reverted.

Also, CudaPerfTests now contains easier method for copying the data
within vectors between host and device.